tests: restructure to test both Tado and TadoX
