+++ test/response/energyIQConsumptionDetails.json
