+++ test/response/eneryIQ.meterReadings.json
