+++ test/response/eneryIQ.meterReadings_update.json
