+++ test/response/mobileDevice.settings.json
