+++ test/response/mobileDeviceGeoLocationConfig.json
