+++ test/response/pushNotificationRegistration.json
