+++ test/response/zone.capabilities.json
